tslint: Extend tslint:recommended
Rather than take a whitelist approach to the rules, I think it would be
better to start with some goal (for example, the tslint:recommended set
of rules) and disable those that we don't follow.  This way, it's easy
to see which ones we could potentially enable.

This doesn't mean that we have to follow all of them.  If there are
rules that we don't want, we can move them to a "Rules from
tslint:recommended that we don't want" section.