Add support for using extended-remote with launch configurations
This extends gdb launch configurations to support extended-remote.
It can be used for remote debugging where code is compiled on host, and
transferred and debugged on target using a remote gdbserver (or something
else speaking GDB/MI).

Note, this is not touching the code added in commit
318ece44cf3a ("Added special commands for extended-remote fix #91"),
as that is only related to gdb attach configuration.