Correct a number of issues related to path substitutions.
LLDB path substitutions were not functional since the path commands
were not being sent to the debugger for any of the configurations.
Additionally, the "settings set target.source-map" command was changed
to "settings append target.source-map" to allow for multiple paths to
be specified.  This command also needed to be sent to the debugger as
a "normal command", as it is not part of the MI.

GDB path substitutions were functional for the launch configuration
and SSH, but were not functional for the attach configuration.
Additionally, the paths were not escaped in the "gdb-set
substitute-path" command, therefore Windows/DOS paths were not
preserved correctly.