Update launch to wait for configuration completion before running.
Removed the timer which was previously used to trigger execution of
the run command in the launch configurations and instead updated the
launch configuration to behave like the attach configuration.  This
now waits until the configurationDoneRequest is received before
sending the run command to the debugger.  This is expected to address
race conditions where breakpoints were not being hit due to the
application running prior to the breakpoints having been configured in
the debugger.