Add support for 'startFrame' parameter in stack trace.
When a stack trace contains many functions calls, the 'startFrame' can
be used in subsequent calls to expand a subset of the complete stack
trace.  Although this parameter is optional, if not utilized, it can
prohibit the interrogation of a long list of stack frames when
combined with the 'levels' parameter.