Implement well defined window for breakpoint configuration.
The previous implementation worked under the assumption that
breakpoint requests would be received from the client prior to the
debugger having completed it's initialization.  However, if the
debugger completed initialization prior to having received breakpoint
requests from the client, those breakpoints requests would be
ignored (i.e., there was a race condition).  Additionally, the
previous implementation "buffered" the receipt of the breakpoint
requests and would only process them after the debugger initialization
had completed (i.e., "debug-ready").

This change delays the sending of the "initialized" event back to the
client (until the debugger truly has completed its initialization).
As a result, there is no longer a need to buffer the breakpoint
requests.  Breakpoints requests are now forced to be received after
the debugger has been initialized.  Furthermore, a previous change
prevented running the application until after the "configurationDone"
request is received.  Therefore, with this change, there is now a well
defined window (between sending the "initialized" event and receiving
the "configurationDone" request), for breakpoint configuration to
occur.