Run npm compilation command in a shell rather than as a process.
The Windows npm command is actually npm.cmd, but this doesn't work
correctly if the command is run as a process from tasks.json, since it
won't properly resolve the npm command to npm.cmd.  This changes the
task to run the command via a shell so that the necessary translation
is made.  This change will works on both Windows and POSIX systems.