Correct additional lint issues detected by ESLint.
The set of ESLint rules will be configured according to the
recommended ESLint rules, with exceptions added to mimic as closely as
possible the current set of TSLint rules that were explicitly
disabled.  The expectation is that the project will move towards using
the recommended rules without exceptions.  Prior to performing the
switch-over to using ESLint, corrections are being performed that
would otherwise be flagged by ESLint.