Remove VSCode dependency from Mocha test harness.
There is no need for VSCode to be used in order to run the current
Mocha tests.  The debug adapter doesn't have a dependency on VSCode
itself, so the harness is reworked to be a standalone Mocha testing
harness.

In addition to the changes to the testing harness, code coverage is
now also collected.  Additional tasks have been added to both run the
test harness as well as collect the coverage metrics.  Furthermore,
coverage metrics are also reported to CodeCov.

Shields have also been added to the README covering continuous
integration and code coverage.  Also the VS Marketplace and Open VSX
badges were also added.