Allow `autorun` commands to complete before continuing or running.
The update makes sure that the `autorun` commands are completed prior
to setting breakpoints initiated from the editor, as well as
performing additional processing related to continuing or running the
application.

Also corrected non-unique error response codes in the GDB launch and
attach and added missing error responses for the LLDB and Mago launch
and attach.