Support unix-based X11 server connections as well as TCP-based ones
Typically, on unix systems, local X11 is connected through the socket `/tmp/.X11-unix/X0`, as indicated by the missing host name part in `DISPLAY=:0` or `DISPLAY=:0.0`. WSLg (the X11 implementation provided with WSL in Microsoft Windows) does not listen on the TCP socket localhost:6000, but on the Unix socket only.

So this commit enables SSH X11 forwarding to connect to the unix domain socket if the X server host name is set to the empty string.